+++ src/data/repositories/SurveyFormD2Repository.ts
@@ -1,33 +1,26 @@
 import { D2Api } from "@eyeseetea/d2-api/2.36";
 import { Future } from "../../domain/entities/generic/Future";
-import { Id, NamedRef } from "../../domain/entities/Ref";
+import { Id } from "../../domain/entities/Ref";
 import { SurveyRepository } from "../../domain/repositories/SurveyRepository";
 import { apiToFuture, FutureData } from "../api-futures";
 import _ from "../../domain/entities/generic/Collection";
 import { ImportStrategy } from "../../domain/entities/Program";
-import { ChildCount, Survey, SURVEY_FORM_TYPES } from "../../domain/entities/Survey";
+import { ChildCount, SURVEY_FORM_TYPES } from "../../domain/entities/Survey";
 import {
-    TrackedEntitiesGetResponse,
     D2TrackedEntityInstanceToPost,
     D2TrackerTrackedEntity,
 } from "@eyeseetea/d2-api/api/trackerTrackedEntities";
 import {
-    getParentDataElementForProgram,
     getSurveyType,
     getTrackedEntityAttributeType,
     isTrackerProgram,
 } from "../utils/surveyProgramHelper";
 import {
-    PREVALENCE_FACILITY_LEVEL_FORM_ID,
-    PPS_WARD_REGISTER_ID,
-    PPS_HOSPITAL_FORM_ID,
     PPS_PATIENT_REGISTER_ID,
     SURVEY_NAME_DATAELEMENT_ID,
     PREVALENCE_SURVEY_NAME_DATAELEMENT_ID,
     AMR_SURVEYS_PREVALENCE_DEA_CUSTOM_AST_GUIDE,
     AMR_SURVEYS_PREVALENCE_DEA_AST_GUIDELINES,
-    PPS_COUNTRY_QUESTIONNAIRE_ID,
-    PREVALENCE_SURVEY_FORM_ID,
 } from "../entities/D2Survey";
 import { ProgramDataElement, ProgramMetadata } from "../entities/D2Program";
 import {
@@ -35,25 +28,11 @@ import {
     mapQuestionnaireToEvent,
     mapQuestionnaireToTrackedEntities,
 } from "../utils/surveyFormMappers";
-import {
-    D2TrackerEntitySelectedPick,
-    mapEventToSurvey,
-    mapTrackedEntityToSurvey,
-    trackedEntityFields,
-} from "../utils/surveyListMappers";
 import { Questionnaire } from "../../domain/entities/Questionnaire/Questionnaire";
 import { ASTGUIDELINE_TYPES } from "../../domain/entities/ASTGuidelines";
 import { getSurveyChildCount } from "../utils/surveyChildCountHelper";
 import { TrackerPostRequest, TrackerPostResponse } from "@eyeseetea/d2-api/api/tracker";
-import { Maybe } from "../../utils/ts-utils";
 import { D2TrackerEvent, D2TrackerEventToPost } from "@eyeseetea/d2-api/api/trackerEvents";
-
-const OU_CHUNK_SIZE = 500;
-
-type Filter = {
-    id: string;
-    value: string;
-};
 
 export class SurveyD2Repository implements SurveyRepository {
     constructor(private api: D2Api) {}
@@ -234,164 +213,6 @@ export class SurveyD2Repository implements SurveyRepository {
                     }
                 }
             );
-        });
-    }
-
-    getSurveys(options: {
-        surveyFormType: SURVEY_FORM_TYPES;
-        programId: Id;
-        parentId?: Id;
-        orgUnitId: Id;
-        chunked: boolean;
-    }): FutureData<Survey[]> {
-        const { surveyFormType, programId, parentId, orgUnitId, chunked = false } = options;
-        const filter = this.getFilterByParentId(programId, parentId);
-
-        return isTrackerProgram(programId)
-            ? this.getTrackerProgramSurveys(surveyFormType, programId, orgUnitId, chunked, filter)
-            : this.getEventProgramSurveys(surveyFormType, programId, orgUnitId, filter);
-    }
-
-    private getFilterByParentId(programId: string, parentId: Maybe<string>): Maybe<Filter> {
-        const filterParentDEId = getParentDataElementForProgram(programId);
-
-        const filter =
-            parentId && filterParentDEId
-                ? {
-                      id: filterParentDEId,
-                      value: parentId,
-                  }
-                : undefined;
-
-        return filter;
-    }
-
-    private getTrackerProgramSurveys(
-        surveyFormType: SURVEY_FORM_TYPES,
-        programId: Id,
-        orgUnitId: Id,
-        chunked = false,
-        filter: Maybe<Filter>
-    ): FutureData<Survey[]> {
-        return chunked
-            ? this.getTrackerProgramSurveysChunked(surveyFormType, programId, orgUnitId, filter)
-            : this.getTrackerProgramSurveysUnchunked(surveyFormType, programId, orgUnitId, filter);
-    }
-
-    private getTrackerProgramSurveysUnchunked(
-        surveyFormType: SURVEY_FORM_TYPES,
-        programId: Id,
-        orgUnitId: Id,
-        filter: Maybe<Filter>
-    ): FutureData<Survey[]> {
-        const ouMode =
-            (orgUnitId !== "" && programId === PREVALENCE_FACILITY_LEVEL_FORM_ID) ||
-            programId === PPS_PATIENT_REGISTER_ID
-                ? "DESCENDANTS"
-                : "SELECTED";
-
-        return apiToFuture(
-            this.api.tracker.trackedEntities.get({
-                fields: trackedEntityFields,
-                program: programId,
-                orgUnit: orgUnitId,
-                ouMode: ouMode,
-                filter: filter ? `${filter.id}:eq:${filter.value}` : undefined,
-            })
-        ).flatMap((trackedEntities: TrackedEntitiesGetResponse<typeof trackedEntityFields>) => {
-            const instances: D2TrackerEntitySelectedPick[] = trackedEntities.instances;
-            return this.getOrgUnitNames(instances.map(instance => instance.orgUnit)).flatMap(
-                orgUnits => {
-                    const surveys = mapTrackedEntityToSurvey(instances, surveyFormType, orgUnits);
-                    return Future.success(surveys);
-                }
-            );
-        });
-    }
-
-    private getTrackerProgramSurveysChunked(
-        surveyFormType: SURVEY_FORM_TYPES,
-        programId: Id,
-        orgUnits: string,
-        filter: Maybe<Filter>
-    ): FutureData<Survey[]> {
-        const orgUnitIds = orgUnits.split(";");
-        const chunkedOUs = _(orgUnitIds).chunk(OU_CHUNK_SIZE).value();
-
-        return Future.sequential(
-            chunkedOUs.flatMap(ouChunk => {
-                return apiToFuture(
-                    this.api.tracker.trackedEntities.get({
-                        fields: trackedEntityFields,
-                        program: programId,
-                        orgUnit: ouChunk.join(";"),
-                        ouMode: "SELECTED",
-                        filter: filter ? `${filter.id}:eq:${filter.value}` : undefined,
-                    })
-                ).flatMap(
-                    (trackedEntities: TrackedEntitiesGetResponse<typeof trackedEntityFields>) => {
-                        const instances: D2TrackerEntitySelectedPick[] = trackedEntities.instances;
-                        return this.getOrgUnitNames(
-                            instances.map(instance => instance.orgUnit)
-                        ).flatMap(orgUnits => {
-                            const surveys = mapTrackedEntityToSurvey(
-                                instances,
-                                surveyFormType,
-                                orgUnits
-                            );
-                            return Future.success(surveys);
-                        });
-                    }
-                );
-            })
-        ).flatMap(listOfSurveys => Future.success(_(listOfSurveys).flatten().value()));
-    }
-
-    private getOrgUnitNames(orgUnitIds: Id[]): FutureData<NamedRef[]> {
-        return apiToFuture(
-            this.api.models.organisationUnits.get({
-                fields: { id: true, name: true },
-                filter: { id: { in: orgUnitIds } },
-                paging: false,
-            })
-        ).flatMap(orgUnitsResponse => {
-            const orgUnits = orgUnitsResponse.objects.map(ou => {
-                return { id: ou.id, name: ou.name };
-            });
-            return Future.success(orgUnits);
-        });
-    }
-
-    private getEventProgramSurveys(
-        surveyFormType: SURVEY_FORM_TYPES,
-        programId: Id,
-        orgUnitId: Id,
-        filter: Maybe<Filter>
-    ): FutureData<Survey[]> {
-        const ouMode =
-            orgUnitId !== "" &&
-            (programId === PPS_WARD_REGISTER_ID ||
-                programId === PPS_HOSPITAL_FORM_ID ||
-                programId === PPS_COUNTRY_QUESTIONNAIRE_ID ||
-                programId === PREVALENCE_SURVEY_FORM_ID)
-                ? "DESCENDANTS"
-                : "SELECTED";
-
-        return apiToFuture(
-            this.api.tracker.events.get({
-                fields: { $all: true },
-                program: programId,
-                orgUnit: orgUnitId,
-                ouMode: ouMode,
-                filter: filter ? `${filter.id}:eq:${filter.value}` : undefined,
-            })
-        ).flatMap(response => {
-            const events = response.instances;
-
-            return this.getOrgUnitNames(events.map(event => event.orgUnit)).flatMap(orgUnits => {
-                const surveys = mapEventToSurvey(events, surveyFormType, programId, orgUnits);
-                return Future.success(surveys);
-            });
         });
     }
