+++ src/webapp/hooks/useSurveys.ts
@@ -2,7 +2,7 @@ import { useCallback, useEffect, useState } from "react";
 import { Survey, SURVEY_FORM_TYPES } from "../../domain/entities/Survey";
 import { useAppContext } from "../contexts/app-context";
 import { useCurrentSurveys } from "../contexts/current-surveys-context";
-import { isPaginatedSurveyList } from "../../domain/utils/PPSProgramsHelper";
+// import { isPaginatedSurveyList } from "../../domain/utils/PPSProgramsHelper";
 import { getUserAccess } from "../../domain/utils/menuHelper";
 import { useCurrentModule } from "../contexts/current-module-context";
 import { GLOBAL_OU_ID } from "../../domain/usecases/SaveFormDataUseCase";
@@ -95,52 +95,35 @@ export function useSurveys(surveyFormType: SURVEY_FORM_TYPES) {
 
         const orgUnitId = getOrgUnitByFormType();
 
-        //Only Patient Forms are paginated.
-        if (isPaginatedSurveyList(surveyFormType)) {
-            compositionRoot.surveys.getPaginatedSurveys
-                .execute(
-                    surveyFormType,
-                    orgUnitId,
-                    parentSurveyId,
-                    currentWardRegister?.id,
-                    currentCaseReportForm?.id,
-                    page,
-                    PAGE_SIZE
-                )
-                .run(
-                    paginatedSurveys => {
-                        setSurveys(paginatedSurveys.objects);
-                        setTotal(paginatedSurveys.pager.total);
-                        setPageSize(paginatedSurveys.pager.pageSize);
-                        setLoadingSurveys(false);
-                    },
-                    err => {
-                        //@ts-ignore
-                        setSurveysError(err.message || err.response?.data.message);
-                        setLoadingSurveys(false);
-                    }
-                );
-        } else {
-            const makeChunkedCall: boolean =
-                surveyFormType === "PrevalenceFacilityLevelForm" && !isAdmin;
-            //Other forms are not paginated.
-            compositionRoot.surveys.getSurveys
-                .execute(surveyFormType, orgUnitId, parentSurveyId, makeChunkedCall)
-                .run(
-                    nonPaginatedSurveys => {
-                        setSurveys(nonPaginatedSurveys);
-                        setLoadingSurveys(false);
-                    },
-                    err => {
-                        //@ts-ignore
-                        setSurveysError(err.message || err?.response.data.message);
-                        setLoadingSurveys(false);
-                    }
-                );
-        }
+        const makeChunkedCall: boolean =
+            surveyFormType === "PrevalenceFacilityLevelForm" && !isAdmin;
+
+        compositionRoot.surveys.getPaginatedSurveys
+            .execute(
+                surveyFormType,
+                orgUnitId,
+                parentSurveyId,
+                currentWardRegister?.id,
+                currentCaseReportForm?.id,
+                page,
+                PAGE_SIZE,
+                makeChunkedCall
+            )
+            .run(
+                paginatedSurveys => {
+                    setSurveys(paginatedSurveys.objects);
+                    setTotal(paginatedSurveys.pager.total);
+                    setPageSize(paginatedSurveys.pager.pageSize);
+                    setLoadingSurveys(false);
+                },
+                err => {
+                    //@ts-ignore
+                    setSurveysError(err.message || err.response?.data.message);
+                    setLoadingSurveys(false);
+                }
+            );
     }, [
         compositionRoot.surveys.getPaginatedSurveys,
-        compositionRoot.surveys.getSurveys,
         surveyFormType,
         currentPPSSurveyForm,
         currentPrevalenceSurveyForm?.id,
